Fix beforeStartHook and afterFinishHook timing
If you have a background for your first scenario the beforeStartHook won't get called correctly. If you filter the `scenarioDefinitions` it might work, but you run into issues with Scenario Outlines not matching as well.

Instead of comparing actual scenario objects, just keep a running count and execure beforeStartHook with 0 and afterFinishHook when it equals the total.